# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: НА_УТРАХ_4 Watch Configurator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /presets/{id}:
    get:
      operationId: getPreset
      tags: [presets]
      summary: Get a single preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Preset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchPreset"
        "404":
          description: Not found
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin/courier)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [payment_pending, cancelled, processing, shipping, arrived]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/my:
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: Get orders for current user by session/telegram
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
        - name: telegramId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin/courier)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: requestOrderCancel
      tags: [orders]
      summary: User requests order cancellation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelRequest"
      responses:
        "200":
          description: Cancel requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/refund:
    post:
      operationId: refundOrder
      tags: [orders]
      summary: Admin refunds an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "200":
          description: Refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    WatchPreset:
      type: object
      required: [id, name, watchfaceGeometry, watchfaceMaterial, watchfaceColor, braceletMaterial, braceletType, braceletColor, handsEnabled, priceStars, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        watchfaceGeometry:
          type: string
          enum: [circle, square, star, drawn]
        watchfaceMaterial:
          type: string
          enum: [plastic, metal]
        watchfaceColor:
          type: string
        watchfaceSize:
          type: string
        braceletMaterial:
          type: string
          enum: [plastic_solid, plastic_segmented, metal_solid, metal_segmented, resin, leather, cotton_fabric]
        braceletType:
          type: string
          enum: [solid, segmented]
        braceletColor:
          type: string
        braceletWidth:
          type: ["string", "null"]
        braceletLength:
          type: ["string", "null"]
        closingMechanism:
          type: ["string", "null"]
        handsStyle:
          type: ["string", "null"]
        handsColor:
          type: ["string", "null"]
        handsEnabled:
          type: boolean
        serialNumber:
          type: ["string", "null"]
        boxType:
          type: ["string", "null"]
        boxColor:
          type: ["string", "null"]
        priceStars:
          type: integer
        createdAt:
          type: string
    WatchConfig:
      type: object
      required: [id, sessionId, watchfaceGeometry, watchfaceMaterial, watchfaceColor, braceletMaterial, braceletType, braceletColor, handsEnabled, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        presetId:
          type: ["integer", "null"]
        watchfaceGeometry:
          type: string
          enum: [circle, square, star, drawn]
        watchfaceMaterial:
          type: string
          enum: [plastic, metal]
        watchfaceColor:
          type: string
        watchfaceSize:
          type: ["string", "null"]
        braceletMaterial:
          type: string
          enum: [plastic_solid, plastic_segmented, metal_solid, metal_segmented, resin, leather, cotton_fabric]
        braceletType:
          type: string
          enum: [solid, segmented]
        braceletColor:
          type: string
        braceletWidth:
          type: ["string", "null"]
        braceletLength:
          type: ["string", "null"]
        closingMechanism:
          type: ["string", "null"]
        handsStyle:
          type: ["string", "null"]
        handsColor:
          type: ["string", "null"]
        handsEnabled:
          type: boolean
        serialNumber:
          type: ["string", "null"]
        customWatchfaceUrl:
          type: ["string", "null"]
        customBackgroundUrl:
          type: ["string", "null"]
        skinFullUrl:
          type: ["string", "null"]
        skinStripeUrl:
          type: ["string", "null"]
        boxType:
          type: ["string", "null"]
        boxColor:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Order:
      type: object
      required: [id, configId, status, totalStars, createdAt, updatedAt]
      properties:
        id:
          type: integer
        configId:
          type: integer
        config:
          $ref: "#/components/schemas/WatchConfig"
        userEmail:
          type: ["string", "null"]
        telegramId:
          type: ["string", "null"]
        telegramUsername:
          type: ["string", "null"]
        sessionId:
          type: ["string", "null"]
        status:
          type: string
          enum: [payment_pending, cancelled, processing, shipping, arrived]
        totalStars:
          type: integer
        cancelComment:
          type: ["string", "null"]
        refundComment:
          type: ["string", "null"]
        courierId:
          type: ["integer", "null"]
        paymentTxId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderList:
      type: object
      required: [orders, total]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
    OrderInput:
      type: object
      required: [configId, totalStars]
      properties:
        configId:
          type: integer
        userEmail:
          type: string
        telegramId:
          type: string
        telegramUsername:
          type: string
        sessionId:
          type: string
        totalStars:
          type: integer
        paymentTxId:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [payment_pending, cancelled, processing, shipping, arrived]
        courierId:
          type: integer
    CancelRequest:
      type: object
      required: [cancelComment]
      properties:
        cancelComment:
          type: string
    RefundRequest:
      type: object
      properties:
        refundComment:
          type: string
...