# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * НА_УТРАХ_4 Watch Configurator API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';
import type { WatchConfig } from './watchConfig';

export interface Order {
  id: number;
  configId: number;
  config?: WatchConfig;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  telegramId?: string | null;
  /** @nullable */
  telegramUsername?: string | null;
  /** @nullable */
  sessionId?: string | null;
  status: OrderStatus;
  totalStars: number;
  /** @nullable */
  cancelComment?: string | null;
  /** @nullable */
  refundComment?: string | null;
  /** @nullable */
  courierId?: number | null;
  /** @nullable */
  paymentTxId?: string | null;
  createdAt: string;
  updatedAt: string;
}